requirements.txt
flask==3.0.0
requests==2.31.0
python-docx==1.1.0
google-auth==2.27.0
google-api-python-client==2.118.0
gunicorn==21.2.0
